code/requirements.txt
pyttsx3
SpeechRecognition
wikipedia
wolframalpha
webbot
pafy
youtube_dl
python-vlc
inquirer
rich
youtube-search-python